self_python/ch_10/requirements.txt
datakart
matplotlib
openpyxl
pandas
python-docx
seaborn
python-dotenv